+++ test/test_ngdc.py
@@ -40,21 +40,21 @@ class NgdcTests(unittest.TestCase):
     # CRR302577 is a small paired-end 16S run (~14 MB total) from CRA004536
     TEST_RUN = 'CRR302577'
 
-    def test_ngdc_ftp_fastq_gz(self):
+    def test_ngdc_http_fastq_gz(self):
         with in_tempdir():
-            extern.run('{} get -r {} -m ngdc-ftp'.format(kingfisher, self.TEST_RUN))
+            extern.run('{} get -r {} -m ngdc-http'.format(kingfisher, self.TEST_RUN))
             self.assertTrue(os.path.getsize('{}_f1.fastq.gz'.format(self.TEST_RUN)) == 6478403)
             self.assertTrue(os.path.getsize('{}_r2.fastq.gz'.format(self.TEST_RUN)) == 8310697)
 
-    def test_ngdc_ftp_with_cra_accession(self):
+    def test_ngdc_http_with_cra_accession(self):
         with in_tempdir():
-            extern.run('{} get -r {} -m ngdc-ftp --cra-accession CRA004536'.format(kingfisher, self.TEST_RUN))
+            extern.run('{} get -r {} -m ngdc-http --cra-accession CRA004536'.format(kingfisher, self.TEST_RUN))
             self.assertTrue(os.path.getsize('{}_f1.fastq.gz'.format(self.TEST_RUN)) == 6478403)
             self.assertTrue(os.path.getsize('{}_r2.fastq.gz'.format(self.TEST_RUN)) == 8310697)
 
     def test_ngdc_ascp_fastq_gz(self):
         with in_tempdir():
-            extern.run('{} get -r {} -m ngdc-ascp ngdc-ftp'.format(kingfisher, self.TEST_RUN))
+            extern.run('{} get -r {} -m ngdc-ascp ngdc-http'.format(kingfisher, self.TEST_RUN))
             self.assertTrue(os.path.getsize('{}_f1.fastq.gz'.format(self.TEST_RUN)) == 6478403)
             self.assertTrue(os.path.getsize('{}_r2.fastq.gz'.format(self.TEST_RUN)) == 8310697)